+++ client/src/components/newsPage/event.js
@@ -7,21 +7,98 @@ import painting2 from "./image/painting2.jpg";
 import painting3 from "./image/painting3.jpg";
 import painting4 from "./image/painting4.jpg";
 
-import aa from "./image/y_a.jpg";
-import bb from "./image/y_b.jpg";
-import cc from "./image/y_c.jpg";
-import dd from "./image/y_d.jpg";
-import ee from "./image/y_e.jpg";
-import ff from "./image/y_f.jpg";
-import xx from "./image/y_xx.jpg";
-import yy from "./image/y_yy.jpg";
-import zz from "./image/y_zz.jpg";
 import { API_URL } from "../config/contansts";
 
-// const images = [painting1, painting2, painting3, painting4];
-// const images = [Painting1, Painting2, Painting3];
-
 function Event(){
+  const slideData = [
+    {
+            imgURL: painting1,
+            created_at: "2023.12.04 ~",
+            title: "랍스터 컬렉션",
+            text: "신선한 리얼 랍스터가 그대로! 올겨울 한정판으로 더욱 특별하게 준비한 써브웨이 랍스터 컬렉션"
+            
+          },
+          {
+            imgURL: painting2,
+            created_at: "2023.12.04 ~",
+            title: "써브웨이 오늘의 수프",
+            text: "오늘의 수프를 원하는 사이즈로! 샌드위치랑 함께 먹으면 든든하고 맛있썹!"
+      
+          },
+          {
+            imgURL: painting3,
+            created_at: "2023.11.01 ~",
+            title: "Eat Fresh Feel Good",
+            text: "써브웨이 X 차은우 신선함 가득! 즐거움 가득! 함께하는 이 순간이 즐거워!"
+      
+          },
+          {
+            imgURL: painting4,
+            created_at: "2020.10.16 ~",
+            title: "말이 안 나올 땐 손으로 주문하자!",
+            text: "말이 안 나올 땐? 손으로 주문하자!"
+      
+          },
+  ];
+  const clone_slide = [slideData[slideData.length-2] ,slideData[slideData.length-1], ...slideData, slideData[0], slideData[1]];
+  const [currentSlide, setCurrentSlide] = useState(2);
+  const [noTransition, setNoTransition] = useState(false);
+
+
+  //다음슬라이드로 이동하는 함수 (마지막 슬라이드에서 첫번째로 이동함) 
+  const nextSlide = () => {
+    if(currentSlide === clone_slide.length - 3){
+      setCurrentSlide((prev) => (prev === clone_slide.length - 1 ? 0 : prev + 1));
+      setTimeout(() => {
+        setNoTransition(true);
+        setCurrentSlide(2);
+      }, 500);
+      setNoTransition(false);
+    } else{
+      setNoTransition(false);
+      setCurrentSlide((prev) => (prev === clone_slide.length - 1 ? 0 : prev + 1));
+    }
+    console.log(currentSlide);
+  };
+  
+
+  // 이전슬라이드로 이동하는 함수
+  const prevSlide = () => {
+    if(currentSlide === 2){
+      setCurrentSlide((prev) => (prev === 0 ? clone_slide.length - 1 : prev - 1));
+      setTimeout(() => {
+        setNoTransition(true);
+        setCurrentSlide(clone_slide.length - 3);
+      }, 500);
+      setNoTransition(false);
+    } else {
+      setNoTransition(false);
+      setCurrentSlide((prev) => (prev === 0 ? clone_slide.length - 1 : prev - 1));
+    }
+    console.log(currentSlide);
+  };
+  
+  // 지정된 인덱스로 이동하는 함수
+  const goToSlide = (index) => {
+      setNoTransition(false);
+      if (index === 1){
+      console.log(index);
+      setTimeout(() => {
+        setCurrentSlide(clone_slide.length - 2);
+      }, 0);
+    }
+    else if(index === clone_slide.length - 2){
+      console.log(index);
+      setTimeout(() => {
+        setCurrentSlide(1);
+      }, 0);
+    }
+    else {
+      console.log(index);
+      setCurrentSlide(index);
+    }
+  };
+
   const [data, setData] = useState([]);
 
   const [select, setSelect] = useState("전체");
@@ -59,10 +136,8 @@ function Event(){
       setIsInitialLoad(false);
     }
   }, [select, isInitialLoad]);
-
   
   // 더보기 함수  
-
 
   const showmore = () => {
     console.log("클릭");
@@ -72,15 +147,12 @@ function Event(){
     })
   };
 
-
-
   return(
     <div id="event_all">
       <div className="y_subtitle">
         <h2> 이벤트ㆍ프로모션 </h2>
       </div>
-
-      {/* <div className="y_painting">
+      <div className="Artist_slide_content">
         <div className="Artist_slide">
           <div className="Artist_slide_viewport" aria-live="polite" >
             <ul className="slide_li" style={{
@@ -90,7 +162,9 @@ function Event(){
               {clone_slide.map((slide, index) => (
                 <li className="slide_li_item">
                   <div key={index} className={`slide_img ${index === currentSlide || (currentSlide === index + slideData.length || currentSlide === index - slideData.length) ? "active" : ""}`} >
-                  <img src={slide.imgURL} alt={`Slide ${index + 1}`} />
+                  <a href="/">
+                    <img src={slide.imgURL} alt={`Slide ${index + 1}`} />
+                  </a>
                   </div>
                   <div className={`slide_info ${index === currentSlide ? "active" : ""}`}>
                     <strong className="slide_info_tit">{slide.title}</strong>
@@ -106,7 +180,6 @@ function Event(){
               <a className="bt_prev" onClick={prevSlide}>Prev</a>
               <a className="bt_next" onClick={nextSlide}>Next</a>
             </div>
-
             <div className="slide_ctl_Dot">
               {clone_slide.map((_, index) => (
                 (index > 1 && index < clone_slide.length - 2) &&(
@@ -123,8 +196,8 @@ function Event(){
             </div>
           </div>
         </div>
-      </div> */}
-
+      </div>
+  
       <div className="y_all_button">
         <div style={{position: "relative"}}>
           <div className="YMJ2_subSelectbar">
@@ -161,136 +234,9 @@ function Event(){
         <div className="y_morebtn">
           <button onClick={showmore}>더보기</button>
         </div>
-
       </div>
-
     </div>
-
-
-)
+  )
 }
+
 export default Event;
-
-{/* <div className="y_list_button">
-  <button className="y_button_l"  onClick={buttonlistClick}>
-    전체
-  </button>
-
-  <button className="y_button"  onClick={buttonlistClick}>
-    진행중인 이벤트
-  </button>
-
-  <button className="y_button_r" onClick={buttonlistClick}>
-    종료된 이벤트
-  </button>
-</div> */}
-
-
-{/* <div className="y_list_item">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={aa}></img>
-    </div>
-    <div className="list_content">
-      <h3>랍스타 컬렉션</h3>
-      <p>2023.12.04 ~ </p>
-    </div>
-  </a>
-</div>
-
-<div className="y_list_item">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={bb}></img>
-    </div>
-    <div className="list_content">
-      <h3>써브웨이 오늘의 수프</h3>
-      <p>2023.12.04 ~ </p>
-    </div>
-  </a>
-</div>
-
-<div className="y_list_item">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={cc}></img>
-    </div>
-    <div className="list_content">
-      <h3>Eat Fresh Feel Good</h3>
-      <p>2023.11.01 ~ </p>
-    </div>
-  </a>
-</div>
-
-<div className="y_list_item">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={dd}></img>
-    </div>
-    <div className="list_content">
-      <h3>스파이시 시리즈</h3>
-      <p>2023.09.01 ~ 2023.12.10</p>
-    </div>
-  </a>
-</div>
-
-<div className="y_list_item">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={ee}></img>
-    </div>
-    <div className="list_content">
-      <h3>고구마 칩 신규 메뉴 출시</h3>
-      <p>2022.10.04 ~ 소진 시</p>
-    </div>
-  </a>
-</div>
-
-<div className="y_list_item">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={ff}></img>
-    </div>
-    <div className="list_content">
-      <h3>말이 안 나올 땐 손으로 주문하자!</h3>
-      <p>2020.10.16 ~ </p>
-    </div>
-  </a>
-</div>
-
-
-<div className="y_list_item hide">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={zz}></img>
-    </div>
-    <div className="list_content">
-      <h3>Lobster Teaser Campaign</h3>
-      <p>이벤트가 종료되었습니다.</p>
-    </div>
-  </a>
-</div>
-
-<div className="y_list_item hide">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={yy}></img>
-    </div>
-    <div className="list_content">
-      <h3>2023 슈퍼팝 프로모션</h3>
-      <p>이벤트가 종료되었습니다.</p>
-    </div>
-  </a>
-</div>
-
-<div className="y_list_item hide">  
-  <a href="#" >
-    <div className="list_photo">
-      <img src={xx}></img>
-    </div>
-    <div className="list_content">
-      <h3>차은우 메뉴 프로모션</h3>
-      <p>이벤트가 종료되었습니다.</p>
-    </div>
-  </a>
-</div>  */}
